Выделил тестовые фикстуры в отдельный conftest.py
