+++ план капкан.txt
@@ -108,10 +108,6 @@ python manage.py dumpdata > all_data.json
 	Установить инструменты для тестирования: pytest, pytest-django, coverage
 	Настроить конфигурацию pytest через pytest.ini
 	Написать тесты для моделей приложения users
-	Написать тесты для моделей приложения posts
-	Проверить создание комментариев и их связь с постами
-	Проверить строковое представление моделей
-	Убедиться в корректности связей между комментариями и постами
   Дневная сессия (4 часа)
 	Написать тесты для API endpoints
 	Протестировать создание, получение и удаление комментариев
